feat(shortcuts): Add separate MACCTRL bit
I'm about to register a Mac shortcut which needs knowledge of whether
Ctrl is pressed. Unfortunately, to match most of the shortcuts, we have
clobbered the Ctrl key with Cmd when we are on a Mac. Having a MACCTRL
bit seemed like the easiest way to be generally right on Cmd/Ctrl but
also break out when we need to.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Ia34c1a60eaa9879131145252c7433dbb1b3e0c55