refactor(keyboard): Move devtools shortcut
Provided it is not handled earlier, a keybind including Ctrl will always
be handled in Map.Keyboard.js where, assuming there is no special case
for it, it will be .preventDefault()ed and sent to core. This poses a
problem for system shortcuts, where we don't actually want to run any
code but instead want the browser or OS to handle the shortcut as normal

Previously the Ctrl+Shift+I shortcut to open devtools was a special case
in Map.Keyboard.js. I'm using the devtools binding throughout as a
test-case for my modifications to the shortcut code. I'm about to
register a lot of system shortcuts that shouldn't be sent to core, and
special casing each of them in Map.Keyboard.js seemed a worse idea than
updating the newer Map.KeyboardShortcuts.ts to support this.

At this stage, the amount of extra support needed is minimal, comprising
only something to avoid running .preventDefault() on the event - we
don't want to *actually* run any code when we get this keybind but by
accepting the keybind in Map.KeyboardShortcuts.ts we would previously
always run .preventDefault(). If we move the preventDefault into the
keybind handling code, we can conditionally run it for most shortcuts,
but allow switching it off on a shortcut-by-shortcut basis 

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Ifdf199c59896d4fda2328f5f644034608b1b8bc5