Calc: Fix view size issue.
Issue: When file/view size is changed, we also check the size of the map. If the file is smaller than the view, we also re-size map and canvas.
While switching between small size (many rows are hidden to bottom) / large size views, we need to check the map's and canvas's sizes. Or the small size is kept for large views.

Signed-off-by: Gökay Şatır <gokaysatir@gmail.com>
Change-Id: Ie6ebbdda8e5b2c78799cae49b627fa476cea8ae2